refactor: discountRate 타입 변경
discountRate에 맞게 더블이 아닌 int 타입으로 변경하여 할인울에 맞게 값 정정

+++ backend/src/main/java/com/sidedish/web/dto/EventCategoryProductDto.java
@@ -13,6 +13,6 @@ public class EventCategoryProductDto {
     private int earlyDelivery;
     private int price;
     private EventBadge eventBadge;
-    private double discountRate;
+    private int discountRate;
     private EventCategory eventCategory;
 }
