+++ backend/src/main/java/com/sidedish/web/dto/MainCategoryProductDto.java
@@ -13,6 +13,6 @@ public class MainCategoryProductDto {
     private int earlyDelivery;
     private int price;
     private EventBadge eventBadge;
-    private double discountRate;
+    private int discountRate;
     private MainCategory mainCategory;
 }
